#2 Hardware Ver.2 Upload
All three boards of Version 2 are uploaded:
- Power Delivery Board ver 1.01
- Microcontroller Board ver 1.01
- Magnetic Encoder Board ver 1.01

I have to see if I need to utilize enable pin of the 3V3 Regulator of the magnetic board to lower power consumption on idle, but for now that is good enough!
I also have to see if I can fit another board in there, cause then we are going for accelerometer and gyro! And that could be the end game for "edge controlling (ahem computing).
But then the microcontroller would probably have been ARM based or equivelant so for now, at this stage of development, these 3 boards should address the major design limitations of version 1.
This issue is now closed!